eval/requirements_eval.txt
ragas>=0.2.0
datasets
pandas
matplotlib
seaborn
litellm
anthropic
requests
python-dotenv
